[refactor] renamed services/config/chuubot to slackbots, code for rtm client should now be compatible with @slack/client v4 syntax
