+++ lib/Slack/Message.js
@@ -1,7 +1,5 @@
 'use strict';
 
-const RTM_MESSAGE_SUBTYPES = require('@slack/client').RTM_MESSAGE_SUBTYPES;
-
 /**
  * Message is a JSON-decoded objects with 7 keys:
  *  - type (usually says "message")
@@ -44,11 +42,11 @@ class Message {
    * @returns {boolean}
    */
   isEditMessage() {
-    return RTM_MESSAGE_SUBTYPES.MESSAGE_CHANGED === this.getSubtype();
+    return 'message_changed' === this.getSubtype();
   }
 
   isDeleteMessage() {
-    return RTM_MESSAGE_SUBTYPES.MESSAGE_DELETED === this.getSubtype();
+    return 'message_deleted' === this.getSubtype();
   }
 
   /**
