+++ lib/Slack/Slackbot.js
@@ -1,9 +1,7 @@
 'use strict';
 
 // Import Slack Client
-const RtmClient = require('@slack/client').RtmClient;
-const CLIENT_EVENTS = require('@slack/client').CLIENT_EVENTS;
-const RTM_EVENTS = require('@slack/client').RTM_EVENTS;
+const { RTMClient } = require('@slack/client');
 const RegexpScript = require('./BotScripts/RegexpScript');
 const Script = require('./BotScripts/Script');
 const Message = require('./Message');
@@ -35,17 +33,17 @@ class Slackbot {
 
   setupRtmConnection() {
     // Setup the client and its listeners
-    this.rtm = new RtmClient(this.bot_token);
-    this.rtm.on(CLIENT_EVENTS.RTM.AUTHENTICATED, () => {
+    this.rtm = new RTMClient(this.bot_token);
+    this.rtm.on('authenticated', () => {
       this.log('Chuubot RTM authentication successful');
     });
-    this.rtm.on(CLIENT_EVENTS.RTM.RTM_CONNECTION_OPENED, () => {
+    this.rtm.on('connected', () => {
       this.log('Chuubot RTM connection established');
       this.is_connected = true;
     });
 
     // Add the base listener framework
-    this.rtm.on(RTM_EVENTS.MESSAGE, this.receiveMessage);
+    this.rtm.on('message', this.receiveMessage);
   }
 
   connect() {
